+++ readthedocs/templates/projects/projects_search_analytics.html
@@ -13,7 +13,7 @@
 {% block project_edit_content %}
   {% if show_analytics %}
 
-    <h3>{% trans "Top queries of all time" %}</h3>
+    <h3>{% trans "Top queries" %}</h3>
     <div class="module-list">
       <div class="module-list-wrapper">
         <ul class="long-list-overflow">
